requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain
langchain-google-genai
langchain-community
